Move HSChain.Control to hschain-control
