+++ hschain-control/HSChain/Control/Util.hs
@@ -7,9 +7,12 @@ module HSChain.Control.Util
   , throwNothingM
   , throwLeft
   , throwLeftM
+    -- * Utils
+  , iterateM
   ) where
 
 import Control.Concurrent.STM
+import Control.Monad
 import Control.Monad.IO.Class
 import Control.Monad.Catch
 
@@ -17,6 +20,8 @@ import Control.Monad.Catch
 atomicallyIO :: MonadIO m => STM a -> m a
 atomicallyIO = liftIO . atomically
 
+iterateM :: (Monad m) => a -> (a -> m a) -> m b
+iterateM x0 f = let loop = f >=> loop in loop x0
 
 ----------------------------------------------------------------
 --
