+++ hschain/hschain.cabal
@@ -64,7 +64,6 @@ Library
                   HSChain.Blockchain.Internal.Engine
                   HSChain.Blockchain.Internal.Engine.Types
                   HSChain.Blockchain.Internal.Types
-                  HSChain.Control
                   HSChain.Crypto.Containers
                   HSChain.Data.CIntMap
                   HSChain.Exceptions
